splitting up block2.rs (#1045)
* splitting up block.rs

* minor modifications

* move `TxInclusionProof` into queryable.rs
